Now prompts to create data folder if it does not already exist. Several bug fixes.
Fixed bug causing crash if window size is not multiple of 2 or 4. Fixed problem where main window and dialogs all appear behind IDE window (PyCharm in this case).

+++ acq4/modules/Patch/PatchWindow.py
@@ -99,7 +99,7 @@ class PatchWindow(Qt.QMainWindow):
             ws = Qt.QByteArray.fromPercentEncoding(six.b(uiState['window']))
             self.restoreState(ws)
             
-        self.ui.splitter_2.setSizes([self.width()/4, self.width()*3./4.])
+        self.ui.splitter_2.setSizes([int(self.width()/4), int(self.width()*3./4.)])
         self.ui.splitter.setStretchFactor(0, 30)
         self.ui.splitter.setStretchFactor(1, 10)
